% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryInput.R
\name{countryInput}
\alias{countryInput}
\title{Custom Country Select Input}
\usage{
countryInput(id, x)
}
\arguments{
\item{id}{The input id of the select Input that will get passed on to the apps'
server.}

\item{x}{A column from a dataframe or tibble that contains country names.}
}
\description{
Creates a Shiny Select Input with custom label style to select a country from
a columns dataset.
}
\details{
The purpose of this function is to provide a quick reference for a common
Shiny Shiny Select Input function that appears multiple times in the app's ui.
}
\examples{
\dontrun{
countryInput(id = "country_cases", x = coronavirus$country)
}





}
\author{
Panagiotis Stylianos
}
