% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidExplorer-package.R
\docType{package}
\name{covidExplorer-package}
\alias{covidExplorer}
\alias{covidExplorer-package}
\title{covidExplorer: Contains a Shiny App with Summaries, Analytics and News related to COVID-19}
\description{
The package deploys a shiny app that allows users to interactively
        browse data regarding the spread of COVID-19 in various countries. The
        purpose of the app is to offer a summary of information regarding 
        COVID-19 with emphasis being given on new cases and latest implemented 
        measures. The datasets are being updated daily. Figures in the Measures tab can be exported as png images.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2020/r-package-assessment-petestylianos.git}
}

}
\author{
\strong{Maintainer}: Panagiotis Stylianos \email{pete.styl@gmail.com}

}
\keyword{internal}
