% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launches the covidExplorer Shiny App}
\usage{
launch_app(ui, server)
}
\arguments{
\item{ui}{a .R script that contains the user interface of the shiny app}

\item{server}{a .R script that contains the app's server}
}
\description{
This function deploys a shiny app that allows users to interactively
browse data regarding the spread of COVID-19 in various countries.
}
\author{
Panagiotis Stylianos
}
