% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_layout.R
\name{my_layout}
\alias{my_layout}
\title{Custom Layout for Shiny's App Ui}
\usage{
my_layout(id, x, output, input_option = NULL)
}
\arguments{
\item{id}{The input id of the select Input that will get passed on to the apps'
server.}

\item{x}{A column from a dataframe or tibble that contains country names.}

\item{output}{The servers output that you want to get displayed}

\item{input_option}{Opional argument to add another select input option for the user.
Defaults to NULL.}
}
\description{
Adds a custom sidebarLayout consisting of a sidebar panel
and a main panel to the Ui when called.
}
\details{
The purpose of this function is to automatically add a custom layout used in all
tabs of the app to remove code duplication and avoid errors.
}
\examples{
\dontrun{
my_layout(id = "country_cases",
x = coronavirus$country,
output = plotOutput(),
input_option = NULL
)
}


}
\author{
Panagiotis Stylianos
}
