% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_theme.R
\name{my_theme}
\alias{my_theme}
\title{Add Custom Theme Plot}
\usage{
my_theme(p)
}
\arguments{
\item{p}{A ggplot2 object}
}
\description{
This function add a custom theme to the server's plots to maintain consistency among figures.
}
\author{
Panagiotis Stylianos
}
