% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\name{spinner}
\alias{spinner}
\title{Custom Spinner}
\usage{
spinner(output)
}
\arguments{
\item{output}{The servers output that you want to get displayed}
}
\description{
Adds a spinner with custom layout to indicate that an output is loading
}
